Rename files and replace std::string with string
