+++ tensorflow/contrib/tensorrt/BUILD
@@ -92,7 +92,8 @@ cc_library(
     ] + if_tensorrt([
         "@local_config_tensorrt//:nv_infer",
     ]) + tf_custom_op_library_additional_deps(),
-    alwayslink = 1,
+    # TODO(laigd)
+    alwayslink = 1,  # buildozer: disable=alwayslink-with-hdrs
 )
 
 tf_gen_op_libs(
@@ -113,7 +114,7 @@ tf_cuda_library(
     deps = [
         "//tensorflow/core:lib_proto_parsing",
     ] + if_tensorrt([
-      "@local_config_tensorrt//:nv_infer",
+        "@local_config_tensorrt//:nv_infer",
     ]),
 )
