+++ tensorflow/contrib/tensorrt/resources/trt_resource_manager.cc
@@ -13,11 +13,11 @@ See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
 
-#include "tensorflow/contrib/tensorrt/resources/TRTResourceManager.h"
+#include "tensorflow/contrib/tensorrt/resources/trt_resource_manager.h"
 #include "tensorflow/core/platform/default/logging.h"
 
 std::shared_ptr<tensorflow::ResourceMgr>
-tensorflow::trt::TRTResourceManager::getManager(const std::string& mgr_name) {
+tensorflow::trt::TRTResourceManager::getManager(const string& mgr_name) {
   // mutex is held for lookup only. Most instantiations where mutex will be held
   // longer will be during op creation and should be ok.
   tensorflow::mutex_lock lock(map_mutex_);
