+++ tensorflow/contrib/tensorrt/resources/trt_resources.h
